comm: fix closing communication channel
There was issue with closing communication channel multiple times.

In socket case, the close function was called even without checking
if the socket is still opened (chack was after this call).

In dbus, there is issue with shared connections - there is no way
to get know that the connection was already dropped. Therefore, we
changing it to the private connection where we have better control
on closing the connection.