ofcds: handle get-config for all datastores
Support also startup and candidate besides the running. There is also
change in case of running, since the capable-switch's id is stored
localy, so the data are returned only in case the id exists, i.e. there
were some edit-config or copy-config that set the id of the controlled
capable-switch. Otherwise, no data from OVSDB are obtained.