ofcds: edit-config for selected items from port and logical-switch
Integrated with the copy-config functions to avoid code duplication.
There are many TODOs for particular configuration elements, but there
is a template how to handle all that edit-config's operations and default
operations.

Currently, only the running datastore is covered. The startup and candidate
will be very similar, but instead of aplying data to OVSDB, the XML tree
will be affected.