of-config: modification of OpenFlow information for port
netconf-tests/openflow_set.sh shows that it is possible to change state
of existing port in OVS.

There is a BUG that must be solved in future: When bridge (or port) is missing
and new one is being created, modification of admin-state etc. of the new port
can't be performed before the whole transaction is commited.
Logically, ioctl() as well as OpenFlow fail because there is no port
(/bridge) to configure and to connect to.