ofcds: fix delete operation in edit-config
particular delete must remove the deleted node from the original document.
This was not done in case of running datastore. It is true, that running
is actually mirrored from OVSDB and it does not have an XML datastore, but
the temporary XML document is used for getting the node where the edit-config
operation is going to be applied, so we have to remove the node also in case
of running datastore to avoid infinite loops.