ofcds: review flow-table management
Do it more simple and avoid some (pseudo)heuristics that actually do
not works.

Fow example, we can not explore the rest of edit xml tree since the
nodes we are looking for could be already processed and removed.

There are still possibility, that in a specially formated edit-config,
things are not going to work, but now the code is less magic.