Revert "of-config: flow-table: fixed remove referenced flow-table"
This reverts commit c63068d6de8476a83b5841a7f55676c6a5611d8f.

Don't do stupid heuristics! Keep the configuration data control to
the client - removing the bridge table's link to an object automatically
when the object is removed is stupid because someone can want to remove
the link manually and then it crashes. OVSDB and OF-CONFIG both have links
between the bridge and ports it uses, so map control of this link directly,
don't be Mr Clippy.