model: remove some conditions from model to allow validation in a real world
The original data-model has some conditions that cannot be acchieved in a real
world. This commit removes (comments) such conditions to allow validation of the
current implementation.

Example: rate element in the features container of the port includes min-elements 1
statement, but there can be a virtual interfaces which actually does not have any
rate. The same case is with the medium element.