getting config data: fix requested-number value
The requested-number is optional element. Neither OVSDB requires it.
However, it was printed everytime - if it was not present, the 0 was
printed. Besides that this behavior is not correct, the 0 value is
not allowed for this parameter so the validation did not pass.

Now, if the requested-number was not specified, it is not returned
by get and get-config.