ofcds: change manipulation with struct smap.
Save some allocation and when changing smap (e.g. external_ids) in
ovsrec structures, use directly the smap in that structures and do
not clone it. It is probably not the clean approach, but it works fine
and saves some memory allocation and deallocation. We can always
change it back to using smap_clone() and smap_destroy().