of-config: propagate error from ofconf_update()
This patch makes it easier to discover bad path to OVSDB and OF sockets.
OF-CONFIG fails on datastore init instead of during copy-config that is
performed later after datastore init. As the effect, error message of
connection failure remains the last one (= it is better to find reason of
error).