of-config: BUGFIX & performance improvement
Rewritten ofc_update() increases performance of OF-CONFIG server, it
doesn't repeat wait when OVSDB replica doesn't differ.

Additionally, this patch fixes problem with start of OF-CONFIG after
crash when OVSDB contains a lot of data.
Crash could be reproduced by creation of 254 flow-tables and attempt to
remove them. During removing OF-CONFIG crashes with "(null): virtual
memory exhausted" receiving SIGABRT.

As a side effect, when OVSDB socket is missing (e.g. because OVS is not
running), OF-CONFIG tries to reconnect infinitely.