ofcds: fix removing conteroller's ip-address
ip-address parameter of the controller is mandatory. However, deleting
can be done as part of the replace/merge operation. But since the value
in OVSDB is mandatory, we cannot set it to NULL, so we are going to
ignore deletion while expecting following creation that actually is going
to really replace the current value.