ofcds: fix removing edit-config's element in remove implementation
After performing edit-config's operation (delete/create), it is needed
to remove the applied element to avoid using it in following (default)
merge. This was done even in remove, but it was done in all cases. However,
in case the deleting the node failed, the node is not supposed to be deleted,
since it is used later when processing the returned error from remove function.