ofcds: fix modification of arrays on OVSDB structures
OVSDB do some magic checks. When we are removing multiple ports
or flow-tables, we are doing it in two steps following the of-config
data model - first, the flow-table itself is removed, in the second
step the link from bridge is removed. But when we update links in
bridge for the first time, OVSDB detects all other removed ports or
flow-tables and update the array of links in a single step. Therefore,
we have to silently ignore if the call to remove flow-table/port link
in bridge detects that there is no such a link to delete. We do checks
for removing non-existing elements earlier, so it is actually ok and
OVSDB IDL do some work instead of us (but it is black magic).