ofcds: detect interface type
OVS has besides the tunnels 2 types of interface - internal and system.
So far we did not set the type, but we should since it should improve
handling the interfaces by OVS. We detect it using ioctl() getting the
driver information - system interface must have the driver and it is not
the "openvswitch".