ofcds: get admin-state from OVSDB record
If the admin-state was not get via OpenFlow (this request is done to
get also other information), we can try to get it directly from the OVSDB
record. So far, we were trying to get it via ioctl, but it does not success
in case it is not the system interface.