ofcds: handle creation of the queue leafrefs in switch
The queue leaf-list element in switch is in fact status, since queues
are connected with the specific port and not with the bridge. So, from
setting only this leafref, it is not possible to link the queue with the
specific port. However, we want to inform user about this. So this commit
adds a check to creation of this leafref and checks what it is trying to
link and according to it it returns error message. In case that the link
between the queue and port is added correctly (by the queue's port element)
in the same edit-config, it silently continues because everything needed
was already done.