ofcds: getting datapath-id of a logical-switch
Setting datapath-id was correctly done via other_config:datapath-id
value in the Bridge table. However, in case this field is not set,
OVS assigns datapath-id automatically, but it is not provided in
the same field, but can be found at datapath_id column of the Bridge
table. Is it not clear, why the column is not marked as status. Now,
we get datapath-id for get/get-config from the mentioned column.

Fixes #11