server: of: try to reconnect to openflow socket
The commit is related to discussion #9 about the connection to OpenFlow
socket issue. When OVS creates new bridge, the new management socket is
opened. It seems that it takes some time until the socket is ready.

ofc-server tries to connect to the socket and if the
connection is 'reset by peer', ofc-server sleeps for 1 second and tries
to connect again. The number of attempts was set to 30.